requirements.txt
# streamlit==1.28.0
# numpy==1.24.3
# pandas==2.0.3
# scikit-learn==1.3.0
# python-Levenshtein==0.21.1
# nltk==3.8.1
# plotly==5.17.0

streamlit==1.28.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
python-Levenshtein==0.21.1
nltk==3.8.1
plotly==5.17.0
sentence-transformers==2.2.2
torch==2.0.1
transformers==4.34.0
scipy==1.10.1